+++ Flow/src/utopia/flow/datastructure/immutable/Pair.scala
@@ -71,6 +71,15 @@ case class Pair[+A](first: A, second: A) extends IndexedSeq[A] with IndexedSeqOp
 	
 	def unary_- = reverse
 	
+	/**
+	  * @return Whether the two values in this pair are equal
+	  */
+	def isSymmetric = first == second
+	/**
+	  * @return Whether the two values in this pair are not equal
+	  */
+	def isNotSymmetric = !isSymmetric
+	
 	
 	// IMPLEMENTED  ----------------------
 	
@@ -230,6 +239,14 @@ case class Pair[+A](first: A, second: A) extends IndexedSeq[A] with IndexedSeqOp
 	  * @return A pair that combines the values of both of these pairs in tuples
 	  */
 	def zip[B](other: Pair[B]) = Pair((first, other.first), (second, other.second))
+	
+	/**
+	  * Compares the two values in this pair using the specified function
+	  * @param f A function for comparing two values with each other
+	  * @tparam B Type of function result
+	  * @return Function result
+	  */
+	def compareWith[B](f: (A, A) => B) = f(first, second)
 	
 	
 	// NESTED   ------------------------------